# spec/contracts/modeling/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: Modeling API
  version: 1.0.0
  description: |
    Синхронный REST API для BC Modeling (docs/backend.md §5).

    Каждая операция ссылается на use case в `x-use-case` и на операцию таблицы
    доступа в `x-operation` (spec/nfr/modeling.yaml#security.authorization): роль
    проверяется на границе use case'а, а не матчером URL. Наполнено на этапе 2
    (2026-10-08) по одному endpoint'у на экспонированный сценарий; до этого
    `paths` был пуст намеренно — endpoint без `UC-*` нарушал бы трассируемость.

    Ошибки — `application/problem+json` (§8.1) с полем `code`: код инварианта
    (`INV-MDL-NNN`) или бизнес-ошибки (`RELATION_NOT_PERMITTED`, `MDL_*`).
    Класс отказа домена определяет статус: не найдено или скрыто списком доступа —
    `404`, конфликт состояния — `409`, нарушение правила — `422`, роль — `403`,
    очищенный снимок — `410`.

    Остаток XML (`raw_xml`, ADR-0017) наружу не отдаётся: это внутренний формат
    interchange.
servers:
  - url: /api/v1
    description: Относительный путь; приложение и фронтенд отдаются одним артефактом.
security:
  - bearerAuth: []
...
paths:
...
  /models/{id}/views:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [views]
      operationId: createView
      summary: Пустая диаграмма ArchiMate в поддереве diagrams
      x-use-case: UC-MDL-002
      x-operation: CreateView
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:     { type: string, maxLength: 500 }
                folderId: { type: string, format: uuid }
      responses:
        '201':
          description: Создано.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ViewSummary' } }
  /views/{id}:
    parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [views]
      operationId: openView
      summary: Payload представления — узлы, рёбра, геометрия (NFR-01)
      x-use-case: UC-MDL-002
      x-operation: OpenView
      responses:
        '200':
          description: Узлы и рёбра в порядке sort_order.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ViewPayload' } }
  /views/{id}/layout:
    parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
    put:
      tags: [views]
      operationId: saveViewLayout
      summary: Геометрия после перемещений одной транзакцией (NFR-03)
      x-use-case: UC-MDL-002
      x-operation: SaveViewLayout
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/LayoutPatch' } }
      responses:
        '204': { description: Сохранено. }
  /views/{id}/nodes:
    parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
    post:
      tags: [views]
      operationId: placeElement
      summary: Разместить элемент; уже размещённый не дублируется (UI-002)
      x-use-case: UC-MDL-002
      x-operation: SaveViewLayout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [elementId, x, y]
              properties:
                elementId: { type: string, format: uuid }
                x:         { type: integer }
                y:         { type: integer }
                width:     { type: integer, description: "По умолчанию 120, как у Archi." }
                height:    { type: integer, description: "По умолчанию 55." }
                parentId:  { type: string, format: uuid, description: "Узел-контейнер; bounds — относительно него." }
      responses:
        '201':
          description: Узел.
          content:
            application/json: { schema: { $ref: '#/components/schemas/ViewNode' } }
  /view-nodes/{id}:
    parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
    delete:
      tags: [views]
      operationId: removeViewNode
      summary: Убрать из представления; элемент остаётся в модели
      x-use-case: UC-MDL-002
      x-operation: SaveViewLayout
      responses:
        '204': { description: Узел, вложенные узлы и их рёбра убраны. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT, выданный Keycloak. Роли VIEWER / ARCHITECT / ADMIN — в claim'ах токена (FR-28), группы — в `groups`.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, maxLength: 64 }
      description: |
        Ключ дедупликации повторов. Тот же ключ с тем же телом возвращает результат
        первого выполнения; тот же ключ с другим телом — `409` (INV-MDL-003).
        Ключ принадлежит пользователю: чужой ключ с тем же значением не конфликтует.
    ModelId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    Property:
      type: object
      properties:
        key:   { type: string, maxLength: 200 }
        value: { type: string }
...
    ViewSummary:
      type: object
      properties:
        id:        { type: string, format: uuid }
        folderId:  { type: string, format: uuid }
        archiId:   { type: string }
        archiType: { type: string }
        name:      { type: string }
        sortOrder: { type: integer, format: int64 }
...
    Style:
      type: object
      description: Собственный стиль объекта — сильнее токена слоя (FR-21). Нет стиля — null.
      properties:
        fillColor:     { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        font:          { type: string }
        fontColor:     { type: string }
        lineColor:     { type: string }
        textAlignment: { type: integer }
    ViewNode:
      type: object
      properties:
        id:        { type: string, format: uuid }
        parentId:  { type: string, format: uuid }
        archiId:   { type: string }
        archiType: { type: string }
        kind:      { type: string, enum: [DIAGRAM_OBJECT, GROUP, NOTE, OTHER] }
        elementId: { type: string, format: uuid }
        x:         { type: integer, description: "Относительно родителя (§3.4, п. 4)." }
        y:         { type: integer }
        width:     { type: integer, description: "-1 — размер по умолчанию Archi." }
        height:    { type: integer }
        style:     { $ref: '#/components/schemas/Style' }
        sortOrder: { type: integer, format: int64 }
    Bendpoint:
      type: object
      properties:
        startX: { type: integer }
        startY: { type: integer }
        endX:   { type: integer }
        endY:   { type: integer }
    ViewEdge:
      type: object
      properties:
        id:             { type: string, format: uuid }
        archiId:        { type: string }
        archiType:      { type: string }
        relationshipId: { type: string, format: uuid }
        sourceId:       { type: string, format: uuid, description: "Узел или ребро." }
        targetId:       { type: string, format: uuid }
        bendpoints:     { type: array, items: { $ref: '#/components/schemas/Bendpoint' } }
        style:          { $ref: '#/components/schemas/Style' }
        sortOrder:      { type: integer, format: int64 }
    ViewPayload:
      type: object
      properties:
        id:            { type: string, format: uuid }
        modelId:       { type: string, format: uuid }
        folderId:      { type: string, format: uuid }
        archiId:       { type: string }
        archiType:     { type: string }
        name:          { type: string }
        documentation: { type: string }
        viewpoint:     { type: string }
        properties:    { type: array, items: { $ref: '#/components/schemas/Property' } }
        editable:      { type: boolean, description: "false у скетча и холста (FR-03)." }
        nodes:         { type: array, items: { $ref: '#/components/schemas/ViewNode' } }
        edges:         { type: array, items: { $ref: '#/components/schemas/ViewEdge' } }
    LayoutPatch:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
            required: [id, x, y, width, height]
            properties:
              id:     { type: string, format: uuid }
              x:      { type: integer }
              y:      { type: integer }
              width:  { type: integer }
              height: { type: integer }
        edges:
          type: array
          items:
            type: object
            required: [id, bendpoints]
            properties:
              id:         { type: string, format: uuid }
              bendpoints: { type: array, items: { $ref: '#/components/schemas/Bendpoint' } }
...